% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Title}
\usage{
add(x, y)
}
\arguments{
\item{x}{First numeric value}

\item{y}{Second numeric value}
}
\value{
Return the sum of the two input values.
}
\description{
Title
}
\examples{
add(3,4)
}
