% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{\subsection{fars_map_state}{
}}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of a state in the US}

\item{year}{a given year}
}
\value{
a plot of accidents' locations of a state
}
\description{
This function prints filename with a given year.
}
\examples{
x <- fars_map_state(26, "2015")
x <- fars_map_state(26, "201"): error because of invalid year
}
