% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{\subsection{fars_read_years}{
}}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
this function returns datasets corresponding to given years,
after adding a year colomn, and selecting month and year colomn
}
\description{
This function read multiple datasets with different given years, and adds a colomn containing the input year.
}
\examples{
x <- make_filename("2015")
x <- make_filename("2015")
x <- make_filename("205"): returns an error because there is not corresponding dataset to year 205

}
