% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{\subsection{fars_read_years}{
}}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years for which number of observations will be counted}
}
\value{
dataframe containing all given years as colomns and a count of observations per month for each year
}
\description{
This function counts number of observations for given years
}
\examples{
x <- fars_summarize_years(c( "2015", "2014)) 
x <- fars_summarize_years(c( "2015", "204)): error because 204 is not valid
}
