% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{\subsection{make_filename}{
}}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year with which  the filename should be created}
}
\value{
prints the created filename
}
\description{
This function prints filename with a given year.
}
\examples{
x <- make_filename("2015")

}
